Adding state logic in main.py
